lab1_Yason_V/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.5.0
pydantic_core==2.14.1
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
fastapi-sso==0.8.0
httpx==0.25.2
python-dotenv==1.0.0
redis==5.0.1  
psycopg2==2.5.0